requirements.txt
streamlit
kagglehub
tensorflow
opencv-python
numpy
pillow
streamlit-webrtc
av
